Core:  - Refactored path manipulators.  - Added frustum-box intersection test.
Engine:
 - Added volume picking.
 - Fixed mesh skeleton debug.
 - Fixed path normalization in mesh texture loading.
 - Fixed orthographic projection near and far planes.

Editor:
 - Added drag selection support.

+++ inc/vapor/animation/Animation.h
@@ -56,7 +56,7 @@ public:
 	bool isLooped();
 
 	// Gets the total animation time.
-	double getTotalTime() const;
+	float getTotalTime() const;
 
 	// Gets the interpolated key frame matrix.
 	Matrix4x3 getKeyFrameMatrix(const BonePtr& bone, double time);
